// types/playerTrophies.ts
export interface PlayerTrophies {
    player_id: string;
    grand_slams: number;        // Títulos de Grand Slam
    masters_1000: number;        // Títulos Masters 1000
    atp_500: number;             // Títulos ATP 500
    atp_250: number;             // Títulos ATP 250
    olympic_golds: number;       // Medallas de oro olímpicas
    olympic_silvers: number;     // Medallas de plata olímpicas
    olympic_bronzes: number;     // Medallas de bronce olímpicas
}